Don't flush the TX buffer right before transmit
This causes random trucated writes, where only 1 or 2 bytes of the packet
are transmitted by the USB serial chip. I'm not exactly sure why; perhaps
there is some kind of buffering that causes the flush command and write
command to make it to the USB device out-of-order.

In any case, I cannot see any reason why a flush should be necessary,
especially on the TX buffer, so it's easy to remove